moved scripts to new home, split search
